backend-api/requirements.txt
# Essential FastAPI dependencies (installed first for better caching)
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic-settings>=2.0.0
pydantic[email]>=2.0.0,<3
sqlalchemy>=2.0.23
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Core async and HTTP dependencies
aiohttp>=3.9.1,<4.0.0
async-generator>=1.10,<2.0
httpx>=0.24.0
requests>=2.31.0
starlette>=0.27.0
uvloop>=0.17.0
python-multipart>=0.0.6

# Data processing and parsing
bs4>=0.0.1,<0.0.2
lxml>=4.9.3,<5.0.0
trafilatura>=1.5.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
python-magic>=0.4.27,<0.5.0
bleach>=6.1.0

# AI and ML core dependencies (grouped for caching)
openai>=1.14.0,<2.0.0
langroid
tiktoken>=0.7.0,<0.8.0

# Document processing
llama-parse>=0.4.1
pypdf>=4.2.0
PyPDF2>=3.0.0
pdfplumber>=0.11.0
pytesseract>=0.3.10
docling
Pillow>=9.0.0

# Data science and numerical computing
numpy>=1.26.0
pandas>=2.0.3,<3.0.0

# Heavy ML dependencies (cached longest)
torch>=2.2.0
sentence-transformers>=2.2.2
onnxruntime>=1.16.1,<2.0.0

# Search and retrieval
qdrant-client>=1.8.0,<2.0.0
rank-bm25>=0.2.2,<0.3.0
whoosh>=2.7.4

# Google services
google-api-python-client>=2.95.0,<3.0.0
google-generativeai>=0.5.2,<0.6.0

# Other AI services
groq>=0.5.0,<0.6.0
grpcio>=1.62.1,<2.0.0

# Utilities and helpers
colorlog>=6.7.0,<7.0.0
docstring-parser>=0.15,<0.16
duckduckgo-search>=6.0.0,<7.0.0
faker>=18.9.0,<19.0.0
fakeredis>=2.12.1,<3.0.0
fire>=0.5.0,<0.6.0
halo>=0.0.31,<0.0.32
jinja2>=3.1.2,<4.0.0
nest-asyncio>=1.6.0,<2.0.0
nltk>=3.8.1,<4.0.0
prettytable>=3.8.0,<4.0.0
pygithub>=1.58.1,<2.0.0
pygments>=2.15.1,<3.0.0
pyparsing>=3.0.9,<4.0.0
pyyaml>=6.0.1,<7.0.0
redis>=5.0.1,<6.0.0
rich>=13.3.4,<14.0.0
slowapi>=0.1.9,<1.0.0
thefuzz>=0.20.0,<0.21.0
typer>=0.9.0,<0.10.0
wget>=3.2,<4.0

# Authentication and security
clerk-backend-api>=3.0.0
PyJWT>=2.6.0
cryptography>=40.0.0
passlib[bcrypt]>=1.7.4
